fixed energize events, making parsing asynchronous for increased performance

+++ WowCombatLogParser/Events/Complex/Suffix/Energize.cs
@@ -5,9 +5,9 @@ namespace WoWCombatLogParser.Events.Complex
 {
     [Suffix("_ENERGIZE")]
     [DebuggerDisplay("{OverEnergize} {PowerType} {AlternatePowerType}")]
-    public class Energize : AdvancedLoggingDetailsBase<int>, IEventSection
+    public class Energize : AdvancedLoggingDetailsBase<decimal>, IEventSection
     {
-        public int OverEnergize { get; set; }
+        public decimal OverEnergize { get; set; }
         public PowerType PowerType { get; set; }
         public PowerType AlternatePowerType { get; set; }
     }
